feat: add Table and CoinTable components
