+++ src/components/coin/table/types.ts
@@ -1,0 +1,12 @@
+export interface ICryptoCoin {
+  id: string
+  name: string
+  image: string
+  current_price: string
+}
+export interface IFirstColumn {
+  id: string
+  image: string
+  value: string
+  path: string
+}
