+++ src/components/shared/table/types.ts
@@ -1,0 +1,6 @@
+export interface ITableData {
+  id: string
+  name: string
+  image?: string
+  current_price?: string
+}
