+++ src/pages/HomePage.tsx
@@ -1,8 +1,8 @@
 import Carousel from '../components/carousel/Carousel'
 import CoinCard from '../components/cards/CoinCard'
-import CoinsTable from '../components/tables/CoinsTable'
 import { useEffect, useState } from 'react'
 import { Grid, Box, Card } from '@mui/material'
+import CoinsTable from '../components/coin/table/CoinsTable'
 
 interface ICryptoCoin {
   id: string
@@ -10,10 +10,10 @@ interface ICryptoCoin {
   image: string
   current_price: string
 }
+
 export default function HomePage() {
   const [data, setData] = useState<ICryptoCoin[]>([])
   const URL: string = `https://api.coingecko.com/api/v3/coins/markets?vs_currency=eur&order=market_cap_desc&per_page=100&page=1&sparkline=false`
-
   useEffect(() => {
     fetch(URL)
       .then(date => date.json())
